refactor: Redis template InMemory interface로 변경
