+++ apps/commerce-api/src/main/java/com/loopers/application/product/ProductFacade.java
@@ -23,6 +23,7 @@ import com.loopers.domain.product.outbox.ProductEventOutbox;
 import com.loopers.domain.product.ProductPayload.ProductEventType;
 import com.loopers.infrastructure.product.ProductWithLikeCount;
 import com.loopers.interfaces.api.product.dto.ProductV1Dto.Response.ProductInfoPageResponse;
+import com.loopers.interfaces.api.product.dto.ProductV1Dto.Response.ProductsInfoResponse;
 
 import lombok.RequiredArgsConstructor;
 
@@ -108,5 +109,9 @@ public class ProductFacade {
         ProductEventOutbox productEventOutbox = productOutboxRegister.register(createProductOutbox);
 
         eventPublisher.publishEvent(new LikeDecrease(productEventOutbox.getId(), productId));
+    }
+
+    public ProductsInfoResponse findProductRanking(String date, Pageable pageable) {
+        return null;
     }
 }
