+++ apps/commerce-api/src/main/java/com/loopers/infrastructure/inmemory/CachedPage.java
@@ -1,4 +1,4 @@
-package com.loopers.infrastructure.redis;
+package com.loopers.infrastructure.inmemory;
 
 import java.util.List;
