+++ apps/commerce-api/src/main/java/com/loopers/infrastructure/inmemory/RedisRepository.java
@@ -1,4 +1,4 @@
-package com.loopers.infrastructure.redis;
+package com.loopers.infrastructure.inmemory;
 
 import java.time.Duration;
 import java.util.Collections;
@@ -10,23 +10,27 @@ import org.springframework.stereotype.Component;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.loopers.application.required.InMemoryRepository;
 
 import lombok.RequiredArgsConstructor;
 
 @Component
 @RequiredArgsConstructor
-public class RedisRepository {
+public class RedisRepository implements InMemoryRepository {
     private final RedisTemplate<String, Object> redisTemplate;
     private final ObjectMapper objectMapper;
 
+    @Override
     public void save(String key, Object value) {
         redisTemplate.opsForValue().set(key, value);
     }
 
+    @Override
     public void save(String key, Object value, Duration ttl) {
         redisTemplate.opsForValue().set(key, value, ttl);
     }
 
+    @Override
     public <T> Optional<T> get(String key, TypeReference<T> typeRef) {
         Object value = redisTemplate.opsForValue().get(key);
         if (value == null) {
@@ -42,10 +46,12 @@ public class RedisRepository {
         return Optional.of(objectMapper.convertValue(value, typeRef));
     }
 
+    @Override
     public void delete(String key) {
         redisTemplate.delete(key);
     }
 
+    @Override
     public <T> List<T> multiGet(List<String> keys, Class<T> type) {
         List<Object> objects = redisTemplate.opsForValue().multiGet(keys);
         if (objects == null || objects.size() == 0) {
